Resolve conflicts in the Plex provider after dev's audiobook/podcast
refactor (get_libraries -> get_section_info/extract_library_name,
logging/dataclass setup):

- helpers.py: keep dev's refactored section helpers and re-add the
  metadata helpers on top (get_explicit, get_musicbrainz_id,
  parse_plex_lyrics_payload + the LRC parsing helpers). Fix the Py3
  syntax error in _lyrics_from_plex_json (except ValueError, TypeError
  -> except (ValueError, TypeError)).
- __init__.py: re-apply mood/style/release_date/explicit/mbid extraction
  in _parse_album/_parse_artist/_parse_track and the lazy lyric fetch in
  get_track on top of dev's parse methods; use contextlib.suppress to
  match the existing style.
- Fix a pre-existing Py3 syntax error carried in from dev's audiobook
  change (#3748) in the album reload handler (except A, B: ->
  except (A, B):) that blocked the module from importing.

+++ music_assistant/providers/spotify/parsers.py
@@ -152,7 +152,7 @@ def parse_track(
         provider=provider.instance_id,
         name=name,
         version=version,
-        duration=track_obj["duration_ms"] / 1000,
+        duration=int(track_obj["duration_ms"] / 1000),
         provider_mappings={
             ProviderMapping(
                 item_id=track_obj["id"],
